Remove calculate_links() from backlink_pages
calculate_links() is designed to be called only once. And the right
place is from scan() function, which scans all the pages. So it
makes no sense to be called from other place. And if it is called,
it will make the result incomplete for entire build.

Currently I founf that backlink_pages can be called by
pagestats::prepprocess even before the scan() is called, which result
in incomplete genration (in my case -> I see no tagcloud)